refactor: refactor ReviewController, move validation to service

+++ src/main/java/ru/yandex/practicum/filmorate/service/ReviewService.java
@@ -8,7 +8,9 @@ import ru.yandex.practicum.filmorate.exception.ReviewNotFoundException;
 import ru.yandex.practicum.filmorate.exception.UserNotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Review;
+import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.review.ReviewDbStorage;
+import ru.yandex.practicum.filmorate.storage.review.ReviewStorage;
 
 import java.util.List;
 import java.util.Optional;
@@ -18,49 +20,81 @@ import java.util.Optional;
 public class ReviewService {
 
     private final ReviewDbStorage reviewDbStorage;
+    private final ReviewStorage storage;
+    private final UserService userService;
 
     @Autowired
-    public ReviewService(ReviewDbStorage reviewDbStorage) {
+    public ReviewService(ReviewDbStorage reviewDbStorage, ReviewStorage storage, UserService userService) {
         this.reviewDbStorage = reviewDbStorage;
+        this.storage = storage;
+        this.userService = userService;
     }
 
-    public Review create(Review review) throws ValidationException {
+    public Review addReview(Review review) throws ValidationException {
         if (isReviewValid(review)) {
             return reviewDbStorage.create(review);
         }
         throw new ValidationException("Validation fail");
     }
 
-    public Optional<Review> update(Review review) throws ValidationException {
-        return reviewDbStorage.update(review);
+    public Optional<Review> updateReview(Review review) throws ValidationException {
+        if (isReviewValid(review)) {
+            return reviewDbStorage.update(review);
+        }
+        throw new ValidationException("Validation fail");
     }
 
-    public void deleteById(long reviewId) {
-        reviewDbStorage.deleteById(reviewId);
+    public boolean deleteById(long reviewId) {
+        return reviewDbStorage.deleteById(reviewId);
+
     }
 
     public Optional<Review> getReviewById(long reviewId) {
-        return reviewDbStorage.findById(reviewId);
+        return storage.findById(reviewId);
     }
 
     public List<Review> getReviewSorted(Long filmId, int count) {
         return reviewDbStorage.getReviewSorted(filmId, count);
     }
 
-    public void addLikeToReview(Long reviewId, Long userId) {
-        reviewDbStorage.addLikeToReview(reviewId, userId);
+    public boolean addLikeToReview(Long reviewId, Long userId) {
+        Optional<User> optUser = userService.findUserById(userId);
+        Optional<Review> optReview = storage.findById(userId);
+
+        if (optUser.isPresent() && optReview.isPresent()) {
+            return storage.addLikeToReview(reviewId, userId);
+        }
+        return false;
     }
 
-    public void addDislikeToReview(Long reviewId, Long userId) {
-        reviewDbStorage.addDislikeToReview(reviewId, userId);
+    public boolean addDislikeToReview(Long reviewId, Long userId) {
+        Optional<User> optUser = userService.findUserById(userId);
+        Optional<Review> optReview = storage.findById(userId);
+
+        if (optUser.isPresent() && optReview.isPresent()) {
+            reviewDbStorage.addDislikeToReview(reviewId, userId);
+        }
+        return false;
     }
 
-    public void deleteLikeFromReview(Long reviewId, Long userId) {
-        reviewDbStorage.deleteLikeFromReview(reviewId, userId);
+    public boolean deleteLikeFromReview(Long reviewId, Long userId) {
+        Optional<User> optUser = userService.findUserById(userId);
+        Optional<Review> optReview = storage.findById(userId);
+
+        if (optUser.isPresent() && optReview.isPresent()) {
+            reviewDbStorage.deleteLikeFromReview(reviewId, userId);
+        }
+        return false;
     }
 
-    public void deleteDislikeFromReview(Long reviewId, Long userId) {
-        reviewDbStorage.deleteDislikeFromReview(reviewId, userId);
+    public boolean deleteDislikeFromReview(Long reviewId, Long userId) {
+        Optional<User> optUser = userService.findUserById(userId);
+        Optional<Review> optReview = storage.findById(userId);
+
+        if (optUser.isPresent() && optReview.isPresent()) {
+            reviewDbStorage.deleteDislikeFromReview(reviewId, userId);
+        }
+        return false;
     }
 
     private boolean isReviewValid(Review review) throws ValidationException, ReviewNotFoundException {
